Fix warning: implicit conversion changes signedness: 'const int' to 'unsigned int' [-Wsign-conversion]
visp/modules/visual_features/include/visp3/visual_features/vpFeatureLuminanceMapping.h:347:16: warning: implicit conversion changes signedness: 'const int' to 'unsigned int' [-Wsign-conversion]
  347 |     m_border = vpFeatureLuminance::DEFAULT_BORDER;
      |              ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ modules/visual_features/include/visp3/visual_features/vpFeatureLuminanceMapping.h
@@ -344,7 +344,7 @@ public:
   void init(const unsigned int k)
   {
     m_mappingSize = k;
-    m_border = vpFeatureLuminance::DEFAULT_BORDER;
+    m_border = static_cast<unsigned int>(vpFeatureLuminance::DEFAULT_BORDER);
     m_Ih = m_Iw = 0;
   }
